refactor KVServiceManager on client: move common code to common super class
